refactor(CRaterRequest): Create AbstractCRaterRequest
